Fixes #20
When trying to re-set the heights of the minimum injections heights
array that correspond to the maximum injection height array that had
been changed, it was mistakenly cropped when also checking which was
greater than a constant value. The two comparisons must be done in one
comparison, not one first then the second.