Fix bug when no events are present in chosen time span
This commit fixes #26, which reported an issue when an array is sent in
to the function `sparse_to_lin` that did not have any events in the
region where the function was asked to interpolate.

The new behaviour is to return empty lists, i.e., no masking, along with
the linspace time axis.