Fix indeterministic test
It was creating the string the file was to be saved as, then the file
was saved (with the name being generated inside the function
independently of the first string). Since `datetime` down to the nearest
second was used, if the file took too long to create, the file names
would not match and the test would fail.

Since only one file is created to the new isolated file system, it
suffices to look for the new file with a glob pattern.