From file (#61)
* First working implementation of eruptions from file

This commit adds the option to load in a file in `json` format, and with
this set the dates and emissions that should be generated.

* Don't use file from repo in tests, generate instead

Removes the use of the `./examples/json.json` file and instead create a
similar one in an isolated file system. Better for testing.

* Remove the shift when generating from file

* Bump version from 0.7.5 to 0.8.0

New version including the `--file` option to read dates and emissions
from a file.