fix(json): allow injections heights to be set alone
The program would error out if only one of minimum or maximum were set
in the json file. This commit allows for that.

Also, if a field in the json file is set but not used/picked up by the
code, a warning message is now printed letting the user know what fields
are still left in the json file. This way it is easier to know you are
using the correct names in the file.